requirements.txt
streamlit==1.25.0
opencv-python==4.8.0.74
mediapipe==0.10.1
Pillow==9.5.0
numpy==1.24.4
